update user panel, homepagesection, return policy

+++ app/Http/Controllers/Admin/BannerController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\Banner;
-use App\Models\bannerImage;
+use App\Models\BannerImage;
 use App\Traits\CreateSlug;
 use Brian2694\Toastr\Facades\Toastr;
 use Illuminate\Contracts\Session\Session;
@@ -25,49 +25,51 @@ class BannerController extends Controller
     // store Banner
     public function store(Request $request)
     {
-        $request->validate([
-            'banner_type' => 'required',
-            'btn_link' => 'required',
-        ]);
 
         $data = new Banner();
         $data->banner_type = $request->banner_type;
         $data->title = $request->title;
-        $data->title_size = $request->title_size;
-        $data->title_color = $request->title_color;
-        $data->title_style = $request->title_style;
+        $data->page_name = $request->page_name;
 
-        $data->subtitle = $request->subtitle;
-        $data->subtitle_size = $request->subtitle_size;
-        $data->subtitle_color = $request->subtitle_color;
-        $data->subtitle_style = $request->subtitle_style;
-
-        $data->text_position = $request->text_position;
         $data->status = ($request->status ? 1 : 0);
-        $data->created_by = Auth::id();
-        $store = $data->save();
+        $data->created_by = Auth::guard('admin')->id();
 
         //if feature image set
-        if ($request->hasFile('phato')) {
-            $allimagepath = [];
-            $images = $request->file('phato');
-            $i = 0;
-            foreach ($images as $image) {
-                $new_image_name = $this->uniqueImagePath('banner_images', 'phato', $image->getClientOriginalName());
+        if ($request->hasFile('banner1')) {
+            $image = $request->file('banner1');
+            $new_image_name = rand(0000, 9999).$image->getClientOriginalName();
+            $image_path = public_path('upload/images/banner/' . $new_image_name);
+            $image_resize = Image::make($image);
+            $image_resize->resize(1170, 250);
+            $image_resize->save($image_path);
 
-                $image_path = public_path('upload/images/banner/' . $new_image_name);
-                $image_resize = Image::make($image);
-                $image_resize->resize($request->width, 250);
-                $image_resize->save($image_path);
-
-                $setBanner = new bannerImage();
-                $setBanner->banner_id = $data->id;
-                $setBanner->phato = $new_image_name;
-                $setBanner->btn_link = $request->btn_link[$i];
-                $setBanner->save();
-                $i++;
-            }
+            $data->banner1 = $new_image_name;
+            $data->btn_link1 = $request->btn_link1;
         }
+        if ($request->hasFile('banner2')) {
+            $image = $request->file('banner2');
+            $new_image_name = rand(0000, 9999).$image->getClientOriginalName();
+            $image_path = public_path('upload/images/banner/' . $new_image_name);
+            $image_resize = Image::make($image);
+            $image_resize->resize($request->width, 250);
+            $image_resize->save($image_path);
+
+            $data->banner2 = $new_image_name;
+            $data->btn_link2 = $request->btn_link2;
+        }
+        if ($request->hasFile('banner3')) {
+            $image = $request->file('banner3');
+            $new_image_name = rand(0000, 9999).$image->getClientOriginalName();
+            $image_path = public_path('upload/images/banner/' . $new_image_name);
+            $image_resize = Image::make($image);
+            $image_resize->resize($request->width, 250);
+            $image_resize->save($image_path);
+
+            $data->banner3 = $new_image_name;
+            $data->btn_link3 = $request->btn_link3;
+        }
+
+        $store = $data->save();
 
         if($store){
             Toastr::success('Banner added successfully.');
@@ -90,47 +92,48 @@ class BannerController extends Controller
      */
     public function update(Request $request)
     {
-        $request->validate([
-            'banner_type' => 'required',
-            'btn_link' => 'required',
-        ]);
-        $data = Banner::find($request->id);
-        $data->banner_type = $request->banner_type;
-        $data->title = $request->title;
-        $data->title_size = $request->title_size;
-        $data->title_color = $request->title_color;
-        $data->title_style = $request->title_style;
 
-        $data->subtitle = $request->subtitle;
-        $data->subtitle_size = $request->subtitle_size;
-        $data->subtitle_color = $request->subtitle_color;
-        $data->subtitle_style = $request->subtitle_style;
-        $data->btn_text = $request->btn_text;
-        $data->btn_link = $request->btn_link;
-        $data->text_position = $request->text_position;
+        $data = Banner::find($request->id);
+        $data->title = $request->title;
+        $data->page_name = $request->page_name;
         $data->status = ($request->status ? 1 : 0);
-        $data->updated_by = Auth::id();
+        $data->btn_link1 = $request->btn_link1;
+        $data->btn_link2 = $request->btn_link2;
+        $data->btn_link3 = $request->btn_link3;
+        //if feature image set
+        if ($request->hasFile('banner1')) {
+            $image = $request->file('banner1');
+            $new_image_name = rand(0000, 9999).$image->getClientOriginalName();
+            $image_path = public_path('upload/images/banner/' . $new_image_name);
+            $image_resize = Image::make($image);
+            $image_resize->resize(1170, 250);
+            $image_resize->save($image_path);
+            $data->banner1 = $new_image_name;
+
+        }
+        if ($request->hasFile('banner2')) {
+            $image = $request->file('banner2');
+            $new_image_name = rand(0000, 9999).$image->getClientOriginalName();
+            $image_path = public_path('upload/images/banner/' . $new_image_name);
+            $image_resize = Image::make($image);
+            $image_resize->resize($request->width, 250);
+            $image_resize->save($image_path);
+            $data->banner2 = $new_image_name;
+
+        }
+        if ($request->hasFile('banner3')) {
+            $image = $request->file('banner3');
+            $new_image_name = rand(0000, 9999).$image->getClientOriginalName();
+            $image_path = public_path('upload/images/banner/' . $new_image_name);
+            $image_resize = Image::make($image);
+            $image_resize->resize($request->width, 250);
+            $image_resize->save($image_path);
+            $data->banner3 = $new_image_name;
+
+        }
+
         $update = $data->save();
 
-        //if feature image set
-        if ($request->hasFile('phato')) {
-            $allimagepath = $data->phato ? json_decode($data->phato) : [];
-            $images = $request->file('phato');
-            foreach ($images as $image) {
-                $new_image_name = $this->uniqueImagePath('banners', 'phato', $image->getClientOriginalName());
-
-                $image_path = public_path('upload/images/banner/' . $new_image_name);
-                $image_resize = Image::make($image);
-                $image_resize->resize($request->width, 250);
-                $image_resize->save($image_path);
-                array_push($allimagepath, $new_image_name);
-            }
-
-            $setBanner = Banner::find($data->id);
-            $setBanner->phato = $allimagepath;
-            $setBanner->btn_link = json_encode($request->btn_link);
-            $setBanner->save();
-        }
 
         if($update){
             Toastr::success('Banner update successfully.');
@@ -143,11 +146,13 @@ class BannerController extends Controller
     public function delete($id)
     {
         $banner = Banner::find($id);
-
         if($banner){
-            $image_path = public_path('upload/images/banner/'. $banner->phato);
-            if(file_exists($image_path)){
-                unlink($image_path);
+            for ($i=1; $i <= $banner->banner_type; $i++) {
+                $banner_image = 'banner'.$i;
+                $image_path = public_path('upload/images/banner/' . $banner->$banner_image);
+                if ($banner->$banner_image && file_exists($image_path)) {
+                    unlink($image_path);
+                }
             }
             $banner->delete();
             $output = [
@@ -164,21 +169,26 @@ class BannerController extends Controller
     }
 
     public function bannerImage_delete(Request $request){
-        $banner = Banner::find($request->id);
-        $allimagepath = $banner->phato ? json_decode($banner->phato) : [];
-        unset($allimagepath[array_search($request->image_path, $allimagepath)]);
+        $bannerImage = Banner::find($request->id);
+        $imageNo = 'banner'.$request->imageNo;
 
-        $image_path = public_path('upload/images/banner/'. $request->image_path);
-        if(file_exists($image_path)){
-            unlink($image_path);
-            //save update path
-            $banner->phato = $allimagepath;
-            $banner->save();
+        if($bannerImage) {
+            $image_path = public_path('upload/images/banner/' . $bannerImage->$imageNo);
+            if ( $bannerImage->$imageNo && file_exists($image_path)) {
+                unlink($image_path);
+            }
+            $bannerImage->$imageNo = null;
+            $bannerImage->save();
+            $output = [
+                'status' => true,
+                'msg' => 'Banner image deleted successfully.'
+            ];
+        }else{
+            $output = [
+                'status' => false,
+                'msg' => 'Banner cann\'t deleted.'
+            ];
         }
-        $output = [
-            'status' => true,
-            'msg' => 'Banner image deleted successfully.'
-        ];
         return response()->json($output);
     }
